Rev 2

+++ app/Http/Controllers/ProductController.php
@@ -2,10 +2,10 @@
 
 namespace App\Http\Controllers;
 
+use App\Http\Resources\ProductResource;
 use Illuminate\Http\Request;
 use App\Models\Product;
 use App\Models\Store;
-use App\Models\User;
 use Illuminate\Support\Facades\Auth;
 use Illuminate\Support\Facades\Storage;
 
@@ -28,17 +28,19 @@ class ProductController extends Controller
         }
 
         // Fetch products belonging to the store
-        $products = Product::where('store_id', $request->store_id)->with('store')->get();
+        $products = Product::where('store_id', $request->store_id)->with(['store', 'category'])->get();
 
-        return response()->json($products);
+        return response()->json(ProductResource::collection($products));
     }
 
 
     public function store(Request $request)
     {
         $request->validate([
+            'category_id' => 'required|exists:categories,id',
             'store_id' => 'required|exists:stores,id',
             'name' => 'required|string|max:255',
+            'measurement' => 'required|string|max:30',
             'price' => 'required|numeric|min:0',
             'quantity' => 'required|integer|min:1',
             'image' => 'nullable|image|mimes:jpeg,png,jpg,gif|max:2048',
@@ -46,7 +48,7 @@ class ProductController extends Controller
 
         $store = Store::find($request->store_id);
 
-        if (Auth::user()->id !== $store->vendor_id) {
+        if ($request->user()->id !== $store->vendor_id) {
             return response()->json(['message' => 'Unauthorized'], 403);
         }
 
@@ -72,7 +74,9 @@ class ProductController extends Controller
         }
 
         $request->validate([
+            'category_id' => 'required|exists:categories,id',
             'name' => 'required|string|max:255',
+            'measurement' => 'required|string|max:30',
             'price' => 'required|numeric|min:0',
             'quantity' => 'required|integer|min:1',
             'image' => 'nullable|image|mimes:jpeg,png,jpg,gif|max:2048',
@@ -94,8 +98,17 @@ class ProductController extends Controller
         return response()->json(['message' => 'Product updated successfully', 'product' => $product], 201);
     }
 
-    public function destroy(Product $product)
+    public function destroy(Request $request, Product $product)
     {
+        $user = $request->user();
+        $product->load('store');
+        $store = $product->store;
+
+        if ($store->vendor_id != $user->id) {
+            return response()->json(['message' => 'Unauthorized'], 403);
+        }
+
+
         if ($product->image) {
             Storage::disk('public')->delete($product->image);
         }
